build pass

+++ tools/util/include/cutlass/util/mixed_dtype_utils.hpp
@@ -132,7 +132,7 @@ CUTLASS_GLOBAL void dequantize_kernel(DequantizedElement* dq_buffer,
       cute::copy(tOpQ_gOpQ(_, _, _, ii), rmem_op_q);
       cute::copy(tScale_gScale(_, _, _, ii), rmem_scale);
       cute::copy(tZero_gZero(_, _, _, ii), rmem_zero);
-      cute::transform(rmem_op_q, rmem_op_scaled, [] (const QuantizedElement& elt) { return ElementScale(elt); } );
+      cute::transform(rmem_op_q, rmem_op_scaled, [] (const QuantizedElement& elt) { return ElementScale((short)elt); } );
       cute::transform(rmem_zero, rmem_zero_buf, [] (const ElementZero& elt) { return ElementScale(elt); } );
       cute::transform(rmem_op_scaled, rmem_scale, rmem_op_scaled, cute::multiplies{});
       cute::transform(rmem_op_scaled, rmem_zero_buf, rmem_op_scaled, cute::plus{});
